soundscape/requirements.txt
dm_haiku==0.0.9
flax==0.6.4
imageio==2.16.1
jax==0.3.21
jax_resnet==0.0.4
librosa==0.9.2
numpy==1.21.2
optax==0.1.4
pandas==1.4.1
PyYAML==6.0
PyYAML==6.0
scipy==1.10.1
soundfile==0.12.1
tensorflow==2.11.0
tqdm==4.64.0
transformers==4.26.1
vit_keras==0.1.2
